Sync with upstream.
Most of these changes are in code we do not use (e.g. the bsd driver) or
for portability.  The key fix we want to pick up is:

    Clear current_bss pointer on disassociation/deauthentication

    This is needed to allow the BSS table entry for the previously used
    BSS to be removed. Now wpa_bss_in_use() can return 0 for the last BSS
    that was used as soon as deauthentication/disassociation event has been
    received.

which manifested itself mostly when roaming after resume (the last AP
would stick in the UI menu).

Review URL: http://codereview.chromium.org/661282