CHROMIUMOS: wpa_supplicant: Add a "flush" command to the "old" wpa_supplicant DBus API
Add a "flush" command to the DBus command per-interface handling.
Revise the innards of wpa_bss_timeout() to break out a function
that can flush all BSSes older than a certain time.  We call the
latter function if a non-zero "age" parameter passed to the "flush"
DBus method, otherwise if no parameter or 0 is passed, we call
wpa_bss_flush() on the interface

TEST=manual test using disappearing base station in WiFi test lab, and http://codereview.chromium.org/3111029
BUG=chromium-os:3157

Review URL: http://codereview.chromium.org/3180020