CHROMIUMOS: wpa_supplicant: Don't send NULL to dbus_message_new_error
The new DBus API helper function wpas_dbus_error_unknown_error
function can be called as a result of a failure within internal
getter calls, which will call this function with a NULL message
parameter.  However, dbus_message_new_error looks very unkindly
(i.e, abort()) on a NULL message, so in this case, we should not
call it.

Getters can be called with NULL by fill_dict_with_properties(), which
iterates through a series of getters.  It doesn't get a message pointer
from the caller, and in the case of signals, none is available.  Furthermore,
it is designed to continue even if there is an error, so that the result
of any successful getters are returned.

I've observed this course of events during a call to
wpas_dbus_getter_bss_wpa with a faileld parse of the IE parameter.
We got here through a call to fill_dict_with_properties which
explicitly calls getters with a NULL message parameter.  Judging
from the way it is called, this could easily occur if an AP sends
out a malformed (or mis-received) probe response.  I usually run
into this problem while driving through San Francisco, so I'm
exposed to any number of base stations along this path.

BUG=chromium-os:7361
TEST=WiFiMatFunc + will do ongoing soak tests during my commute

Review URL: http://codereview.chromium.org/3590019