CHROMIUMOS: wpa_supplicant: Accept CQM bitrate events
Unify the RSSI and txrate state retrieval code.  Add calls for asking
the kernel for txrate CQM events and enhance the bgscan code to be able
to accept CQM events that may or may not contain RSSI threshold
information (in the case where the CQM event was caused by a bitrate
change instead).

BUG=none
TEST=Manual -- turn up debugging and set a bitrate threshold -- and do
walkaround

Review URL: http://codereview.chromium.org/3795003