CHROMIUMOS: wpa_supplicant: Add central "bgscan_monitor" functionality with noise floor calibration
Add a bgscan-core capability for enabling and disabling connection
monitoring (was scattered among bgscan modules) with noise-floor
tracking (new feature).  This means that a desired RSSI value specified
in a bgscan config will be tracked against received noisefloor values.
Although this code is meant to be central to bgscan and used by all
modules, we are staging this new functionality, and will temporarily
create a "bgscan_delta" background scan algorithm which differs from
"bgscan_simple" only in the fact that it uses this new facility.

BUG=chromium-os:8312
TEST=Walk-around test

Review URL: http://codereview.chromium.org/4167005