CHROMIUMOS: wpa_supplicant: Back off short scan interval after a number of attempts
The short scan interval is engaged when the received signal intensity
is (or becomes) lower than our threshold.  It is used to quickly scan
to find an available place to roam to.  This CL adds the behavior of
"giving up" on a quick roam after a few scan attempts, and reverting
to the slow scan interval.  This slows down scans again and preserves
battery in situations where there all the options are bad (or where
there is only one AP and you are stationary in a bad reception zone).

Sameer, you may want to test this change as well.

BUG=chromium-os:6408
TEST=Manual -- join a singleton access point, go to the edge of the
network, watch /var/log/messages and confirm scans.

Review URL: http://codereview.chromium.org/4142013