CHROMIUMOS: wpa_supplicant: add fast reconnect support
When dropped by AP for inactivity or because the session has timed
out, try to re-join directly before falling back to a full scan and the
normal work.  This reduces the time to re-join a network from 5-10seconds
to ~300ms (not counting any time to renew DHCP).

Note that for mac80211 we cannot directly re-authenticate because by the
time we issue a request mac80211 has discarded the BSS state which is
required to process an MLME request.  Instead we scan the channel where
we last saw the AP and then use the normal join mechanism.  Since this
is not true for all drivers it might be worthwhile to try a direct Auth
first and only then fall back to scanning.  Also, note that if the scan
request collides with another scan then we just piggyback on that scan.

Change-Id: I31c869c565e36cc41f7698cad2a3e0a455316eeb