wpa_supplicant: ignore fast reconnect scan request failures
This is a temporary hack to get fast reconnect being used more often.
When fast reconnect kicks in there may be a bgscan active and/or the
system may choose to scan.  In those cases we were falling back to
the slow path.  Instead just assume the scan will complete and we'll
be notified which will result in the fast reconnect logic being run.
This reliably fixes fast reconnect for several scenario's I've hit.

The problem with this hack is the scan request might fail for other
reasons.  To handle this we probably want to change the calls to return
an error code so we can identify when we collided with an outstanding
scan request.  Note this does not appear to be an issue for the devices
we presently care about.

BUG=chromium-os:13075
TEST=manual:connect to ap, suspend long enough for ap to drop sta for inactivity; resume and check logs for fast reconnect failure/success

Review URL: http://codereview.chromium.org/6688046

Change-Id: I19830d5d2ade771f23afae0ce44b2cada3205d67