CHROMIUMOS: bgscan_simple: Time out short scan interval
We gain diminishing returns by the short scan interval.  The short
scan interval is used to hunt for a better AP if the RSSI of the
current AP drops.  However, if we never roam, and the AP continues
to have low RSSI, at some point we should give up and return to
the slow background scan rate, otherwise we waste a lot of power.

BUG=chromium-os:8312
TEST=Manual -- monitor messages from wpa_supplicant

Change-Id: I014db9f66a58c76869123f866f4f787283ff65c6