CHROMIUMOS: bgscan: Add centralized signal_monitor
The signal strength number tells only half of the picture with
respect to signal quality.  The other half is the prevailing
noise on the channel.  The two together indicate how easily the
interface is able to pull data out of the background noise.

This change takes noise into account, by caclulating a CQM
threshold based both on the RSSI and noise.  This noise value is
available in the CQM events as well as polled data.  Since most
background scan algorithms will probably want to do this the same
way, I have added this as routines in bgscan.c that can be
utilized by individual background scan mechanisms.

Additionally, I've modified bgscan_simple to use this mechanism,
and scheduled noise floor updates to happen at the same rate as
background scans (the noisefloor should be a slowly changing
parameter).

BUG=chromium:190663
TEST=Manual -- observe nl80211 messages within supplicant indicating
     actual CQM thresholds, as well as timing between these and bgscan
     messages indicating updates.

Change-Id: I94501dd21b0b2ca9142902097c27c8b31d0c17f4
Original-Reviewed-on: https://chromium-review.googlesource.com/323774
Original-Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Original-Tested-by: Eric Caruso <ejcaruso@chromium.org>
Original-Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/669502
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>