UPSTREAM: Drop oldest unknown BSS table entries first
When the BSS table size limit has been reached, drop first the oldest
BSS entries for which there is not a matching network in the
configuration based on SSID (wildcards are ignored). This makes it
less likely to hit connection issues in environments with huge number
of visible APs.

BUG=chromium-os:20486
TEST=Manual (can't generate enough BSSes in the testbed): Ensure scans
still work as before.  Use dbus-send to a wpa_supplicant modified with
http://gerrit.chromium.org/gerrit/7826 to set max_bss below the number
of BSSes that support the current SSID I'm connected to.  Note that the
nasty churn from chromium-os:20486 shows up immediately as the connected
BSS gets removed.  Reset max_bss just above the number of BSSes for my
SSID that are visible, and the churn stops by itself without restarting
anything.

Change-Id: I9cc563d2932abc96f7249134da5c4ebd3da8a553
Reviewed-on: http://gerrit.chromium.org/gerrit/7811
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>