wpa_supplicant: Prefer 5 GHz networks over 2.4 GHz networks.
In scan.c, merge a channel's noise value into the scan results.  When comparing
scan results, compute the signal-to-noise ratio and use it when available.
Prefer a 5 GHz network if its SNR is really big (> 30) or if its SNR is
relatively close to the other network's.

BUG=chromium-os:18867
TEST=Added new auto-test WiFiRoaming:008Prefer5GHz.  Also, verified on
GoogleGuest that 5GHz channel is preferred.  Also ran MatFunc and SecMat.

Change-Id: I804a6ac79410633362fad91cd1ce8c3570a6e64a
Reviewed-on: http://gerrit.chromium.org/gerrit/5576
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>