wpa_supplicant: Modify bgscan based roaming behaviour.
Cancel background scan based roaming behavior if the driver
is capable of handling roaming in the firmware. For fullmac
drivers like ath6kl, the roaming logic is handled in the
firmware and for supplicant to initiate roam, the station
has to disconnect completely and then reconnect with the
new AP which takes a longer time without preauth.

Currently, if the supplicant also tries to roam, all requests
are rejected by cfg80211 resulting in a spam of log.

Signed-off-by: Vivek Natarajan <nataraja@qca.qualcomm.com>

BUG=chrome-os-partner:5421
TEST=manual:site walk around while associated; check logs to verify roaming is supppressed

Change-Id: I5cb89bd0c377e1a4602890a36955430d8487b691
Reviewed-on: http://gerrit.chromium.org/gerrit/7007
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>