CHROMIUMOS: wpa_supplicant: filter scan results during fast reconnect
When doing fast reconnect the one channel scan we do to setup a kernel
bss entry (so we can join the network) may return cached data.  If this
data includes the bss we want to reconnect to then we may try to re-join
it even though it's not there.  To work around this filter scan results
using the "last seen" timestamp on scan records--if the entry was not
seen during our last scan toss it.  This may cause us to do more work
if we miss the AP in the scan but otherwise have it cache'd but that's
better (i.e. faster) than using stale data.

BUG=chromium-os:15981
TEST=006BeaconLoss from Roaming suite

Change-Id: I274c956019409ef21656b0d6b0d42f18ddaab7f9
Reviewed-on: http://gerrit.chromium.org/gerrit/7437
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>