CHROMIUMOS: wpa_supplicant: handle auth failure during fast reconnect
When doing fast reconnect the auth request may fail for a variety of
reasons.  Carry the fast reconnect state forward through auth so if
it fails we can properly cleanup the state machine and notify clients
about losing connectivity.  We do this only for the SME case as it
is the only one that can currently implement fast reconnect.

BUG=chromium-os:15981
TEST=006BeaconLoss from the WiFiRoaming suite

Change-Id: I3839ed712a89606f23069fab1dc730ff90bf783e
Reviewed-on: http://gerrit.chromium.org/gerrit/7438
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>