CHROMIUMOS: wpa_supplicant: Refinements to fast-scan backoff
These changes account for situations where the CQM threshold
might be approximately the same as the currently received signal,
and thus CQM events are triggered often due to measurement
error/small fluctuations.  This results in scanning occurring
too frequently.

Firstly, inhibit the immediate scan when the short-scan count
is at the maximum.  This keeps bursts of CQM toggling from
causing a torrent of back-to-back scans.  This does not inhibit
immediate scans if the CQM triggers a second time (if the signal
falls lower past the hysteresis).  This reduces the scan rate in
the worst case (fast-rate toggling high/low CQM events) to the
short scan interval.

Secondly, change the behavior of the short scan count so it acts like
a "leaky bucket".  As we perform short-scans, the bucket fills until
it reaches a maximal short-scan count, at which we back-off and
revert to a long scan interval.  The short scan count decreases by
one (emptying the bucket) every time we complete a long scan interval
without a low-RSSI CQM event.

This reduces the impact of medium-rate toggling of high/low CQM
events, reducing the number of short-interval scans that occur before
returning to a long-interval if the system was recently doing
short scans.

BUG=chromium-os:21371
TEST=Manual -- set threshold and experiment with crossing backwards
and forwards across CQM boundary.  To do this, observe the current
signal and noise values ("wpa_debug msgdump", then wait for "Signal"
and "Noise" messages in /var/log/messages).  You should then set your
threshold to "-95 + Signal - Noise", using the command:
   /usr/local/lib/flimflam/test/set-bgscan BgscanSignalThreshold=<threshold>
At this point, simply moving closer or futher from your AP should
cause CQM events.

Change-Id: I6bf38d6aad4be99fbe8ca6964631b53ccfbd8c92
Reviewed-on: http://gerrit.chromium.org/gerrit/9711
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>