CHROMIUMOS: wpa_supplicant: improve roaming when noise floor available
When the noise floor is available we can do a better job deciding when
to roam by using SNR instead of the signal level.  In addition roam only
if the current AP's snr is below a threshold.  For situations where the
noise floor is not available we fall back to the previous algorithm.
New config parameters roam_threshold and roam_min hold the values used
by the roaming algorithm. These can also be set through the control
interface.

BUG=chromium:201970
TEST=WiFiRoaming suite + walk-around tests + ATEN roaming tests

Change-Id: I6f7c16ec2eac3660dfbec22bb55b82daae11fbe2
Original-reviewed-on: https://gerrit.chromium.org/gerrit/11982
Original-reviewed-by: Sam Leffler <sleffler@chromium.org>
Original-tested-by: Sam Leffler <sleffler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/669507
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>