CHROMIUMOS: wpa_supplicant: improve roaming when noise floor available
When the noise floor is available we can do a better job deciding when
to roam by using SNR instead of the signal level.  In addition roam only
if the current AP's snr is below a threshold.  For situations where the
noise floor is not available we fall back to the previous algorithm.
New config parameters roam_threshold and roam_min hold the values used
by the roaming algorithm. These can also be set through the control
interface.

BUG=chromium-os:18867
TEST=WiFiRoaming suite + walk-around tests + ATEN roaming tests

Change-Id: I1aae8c492618d01242462c2ee76b76fd8a9df0a3
Reviewed-on: https://gerrit.chromium.org/gerrit/11982
Commit-Ready: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>