UPSTREAM: nl80211: Filter unexpected interface added/up events
It looks like a RTM_NEWLINK event claiming the interface to be UP is
delivered just before removing an interface after having first indicated
that the interface was going down/removed. Ignore this event if the
interface is not present anymore at the moment the event is processed.
This fixes issues where an interface that was re-added after being
removed did not get reconfigured properly.

Signed-hostap: Jouni Malinen <j@w1.fi>
intended-for: hostap-1
(cherry picked from commit d1f4942ba18a303d275bbd955161674939d1d902)

BUG=chromium:434785
TEST=On a WP2-equipped device execute "rmmod iwlmvm; modprobe iwlmvm",
and ensure that the network dropdown continues to show a list of
available networks and that the system reconnects (if it was connected
before).,

Change-Id: I5e4d637300ef06047331a65133f1eef91c2830df
Reviewed-on: https://chromium-review.googlesource.com/230810
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>