CHROMIUM: wpa_supplicant: export disconnect reason code to dbus
In the properties changed signal, added a new propery "DisconnectReason",
which carries the IEEE 802.11 reason code of the most recent
disassociation or deauthentication event.  The reason code is negative
if it is locally generated.  The property is sent to the DBUS
immediately so as to prevent it from being coalesced with other
disconnect events.

BUG=chromium-os:26854
TEST=Manual testing.  Disabled WiFi interface and, using dbus-monitor, saw
DisconenctReason == -3, which means DEAUTH_LEAVING, locally generated.

Change-Id: Id1c3f2b63422546ead00f9649b24bd97a73696d1
Reviewed-on: https://gerrit.chromium.org/gerrit/16740
Commit-Ready: Gary Morain <gmorain@chromium.org>
Reviewed-by: Gary Morain <gmorain@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>