CHROMIUMOS: wpa_supplicant: Don't leak scan frequencies
Use a single buffer for storing the list of frequencies to use
for background scan (in the order they should be scanned).
Instead of allocating it once for each scan (then leaking it,
as the previous implementation did), create the buffer at
the start of the background scan method.

BUG=chromium-os:30050
TEST=valgrind

Change-Id: I743caecde13ffd973262d51b336c255af62835af
Reviewed-on: https://gerrit.chromium.org/gerrit/24829
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>