CHROMIUM: Don't blacklist SSID for locally generated disassociation
During fast assocation process, trying to associate to AP1, we disconnect with
currently connected AP2. During this process, a locally generated
disassociation event happens resulting in AP2 sending a dissassociation frame.
wpa_supplicant wrongly thinks that it got a disassociation event from AP1 and sends AP1
to blacklist. So fix this by checking if the disassociation is generated
locally or not.

BUG=chromium:220857
TEST=Manual( 1. Turn wpa debug to excessive
	     2. Using script or manually, associate alternatively between two
	     different APs repeatedly
	     3. Shouldn't see any of these APs added
	     to blacklist for locally generated disassociation)

Change-Id: Ibf64cfa65394addb54054ce7ff5f983be8d9fefc
Signed-off-by: Ashok Nagarajan <asnagarajan@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47263
Reviewed-by: Paul Stewart <pstew@chromium.org>