CHROMIUM: hostapd: Provide compile option for weak random numbers
Provide a configuration option for allowing the use of /dev/urandom
instead of /dev/random for the random number generator source.
This source performs more quickly at the expense of security but
can be useful in some limited scenarios like a testbed for WPA
authentication where the the client should not be denied access
just because the random number pool is too small.

BUG=chromium:218911
TEST=Compile with and without use-flag; strings the executable

Change-Id: I2b92788205450e3c51dbb725f292d3495f615433
Reviewed-on: https://gerrit.chromium.org/gerrit/48906
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>