Improve performance of reconnect operation
Added "reattach" command to perform single-channel single-ssid
scan instead of full scan when trying to reconnect to currently
"connected" network. This allows the scan result to come
back in much faster time. In ath9k, the scan took around 12 seconds
with full background scan, and only 0.1 second with the single-channel
single-ssid scan.

Since for reconnect operation the only SSID we care about is
the one we previously connected to, so it is good idea to
perform single-channel single-ssid scan.

BUG=chromium:264865
TEST=manual
     1. Verify network_WiFi_Reassociate test passes
     2. Run full matfunc test suite

Change-Id: I6f65111c9a05ee2d7657a82794bed9fa89e8a90a
Reviewed-on: https://chromium-review.googlesource.com/189051
Tested-by: Peter Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>