CHROMIUM: wpa_supplicant: fix nl80211_toggle_high_bitrates
When disabling high bitrates, we ask the kernel to enable (only)
the 1, 2, 5.5, 6, and 11 Mbps. We do this for both the 2.4 GHz
band, and the 5 GHz band.

Unfortunately, most of these bitrates are invalid for the 5 GHz
band. Furthermore, newer kernels (including 3.8) check for invalid
bitrates, and reject the configuration request entirely if any
are found.

Resolve this problem by configuring the 5 GHz band for legacy
bitrates of 6, 9, and 12 Mbps. These are the three lowest
bitrates reported by "iw phy phy0 info", for the 5 GHz band.

There was no problem with the MCS configuration, so that is
left alone. That is, we still enable MCS 0 and MCS 1 in both
bands.

BUG=chromium:350275
TEST=network_WiFi_LowInitialBitrates on peach_pi

Change-Id: I880ebde0080cfc1f2c59eec9a31416ece9973fc5
Reviewed-on: https://chromium-review.googlesource.com/189148
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>