wpa_supplicant: reduce overall time takes for peering.
The beacon interval is 1 second. it takes around 3 seconds for the HW to start
beaconing. So totally it takes 4 seconds to discover and peer with new mesh
point. Reduce this time to 0.5 seconds by 1) sending a broadcast probe request
as soon as the mesh interface is up 2) sending directed probe request as soon
as an auth frame is received from an unknown sta.

BUG=b:27324978
TEST=Verified by connecting 2 Mesh Points (root and non root) in a loop.

Change-Id: I748e4f24ecc040fd02d0cdfdcd06aa504a380657
Reviewed-on: https://chromium-review.googlesource.com/339830
Commit-Ready: Srinivasa duvvuri <sduvvuri@chromium.org>
Tested-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: John Walker <walker@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>