UPSTREAM: OpenSSL: Initialise PKCS#11 engine even if found with ENGINE_by_id()
Recent versions of engine_pkcs11 are set up to be autoloaded on demand
with ENGINE_by_id() because they don't need explicit configuration.

But if we *do* want to explicitly configure them with a PKCS#11 module
path, we should still do so.

We can't tell whether it was already initialised, but it's harmless to
repeat the MODULE_PATH command if it was.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>
Tested-by: Michael Schaller <misch@google.com>
(cherry picked from commit c3d7fb7e2724150950e1a1eac29460ea255811c3)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:566283
TEST=network_WiFi_SimpleConnect.wifi_check1x_WPA

Change-Id: I31a58d28f9703b853713401289df5f3f4cbad747
Reviewed-on: https://chromium-review.googlesource.com/571540
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>