UPSTREAM: utils: Add ssid_parse() function
Add a function that parses SSID in text or hex format. In case of the
text format, the SSID is enclosed in double quotes. In case of the hex
format, the SSID must include only hex digits and not be enclosed in
double quotes. The input string may include other arguments after the
SSID.

[cherry pick from 624b8a061f502d09ba932984482195bd5a839db5]
[git remote add hostap git://w1.fi/srv/git/hostap.git]

BUG=b:36055223
TEST=1) Manually verified the neighbor report functionality by sending a
     neighbor report request from a STA and receiving response from AP.
     2) stress tested by running through  connect/disconnect/download
        from 7 different stations and observed no regression.

Change-Id: I498108aeb0b6dffea89bf9e332683b4923ec35be
Reviewed-on: https://chromium-review.googlesource.com/456536
Commit-Ready: NARAYANRADDI MASTI <nmasti@google.com>
Tested-by: NARAYANRADDI MASTI <nmasti@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>