UPSTREAM: Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Change-Id: I82517039bb846bb6ecde4e00dbe91a4be1b3ec19
Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>
Reviewed-on: https://chromium-review.googlesource.com/721270
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>