UPSTREAM: FT: Do not allow multiple Reassociation Response frames
The driver is expected to not report a second association event without
the station having explicitly request a new association. As such, this
case should not be reachable. However, since reconfiguring the same
pairwise or group keys to the driver could result in nonce reuse issues,
be extra careful here and do an additional state check to avoid this
even if the local driver ends up somehow accepting an unexpected
Reassociation Response frame.

Change-Id: I95736fc5a53127320dc9ea6634df49fc33caf953
Signed-off-by: Jouni Malinen <j@w1.fi>
Reviewed-on: https://chromium-review.googlesource.com/721276
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>