CHROMIUM: wpa_supplicant: Full-scan after blacklisting
wpa_supplicant currently blacklists APs that deauth it, and
restricts the next scan to frequencies it has seen in prior scans
of the ESS it was just connected to. This precludes APs that have
changed frequencies from being discovered on subsequent scans.
See CL:421224 for more details.

This change adds a config to disable the scan restriction to
previously seen frequencies.

BUG=chromium:675232
TEST=tested manually

Change-Id: I3b127a29d5a6078457bf111419c034ced3bcff06
Reviewed-on: https://chromium-review.googlesource.com/740821
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>