hostapd:[ConfigKnob] Return SUCCESS on deleting an entry
Current behaviour:
Current implementation returns FAIL if the sta_id is not
available as part of the sta_policy.conf, Due to which
the entries on the upper layer where not deleted.

Changes made:
Returning SUCCESS even if the sta_id is not available
on the station list. Which would allow the upper layer
to remove the STA entry

BUG=b:73220508
TEST=Below are the observations.
1.Try deleting a sta_id entry which is not present
   in the list.
   Output : Should return SUCCESS.
2. Keep the sta_policy list empty and try deleting an
   entry.
   Output : Should return SUCCESS.

Change-Id: I8eef63a2b523631070187f0fb2d884cc5fa72a0c
Reviewed-on: https://chromium-review.googlesource.com/1095197
Commit-Ready: Bhogeswara Rao Metta <bhogeswararao@google.com>
Tested-by: Bhogeswara Rao Metta <bhogeswararao@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@google.com>
Reviewed-by: ratheesh.sp <ratheesh.sp@globaledgesoftware.corp-partner.google.com>