wpa_supplicant: fix the constant roaming problem
We saw that on certain platforms in certain places we keep
switching between two APs and eventually get the same RSSI.
Debugging showed that we have a very big difference between
the two antennas.

Ant A can hear AP A very well (-60) but AP B very bad (-80)
Ant B can hear AP B very well (-60) but AP A very bad (-80)

When the device associates to AP A, it'll learn to use Ant A.
If the device uses one single antenna to receive the scan
results, it may hear the AP it is currently associated to
on the second antenna and get bad results. Because of that,
the wpa_supplicant will roam to the other AP and the same
scenario will repeat itself:

Association to AP A (Ant A reports -60).
Scan on Ant A: AP A: -60, AP B: -80
Scan on Ant B: AP A: -80, AP A: -60 ==> ROAM.

Association to AP B (Ant B reports -60)
Scan on Ant A: AP A: -60, AP B: -80 ==> ROAM

Etc...

Fix this by querying the signal of the current AP using
drv_signal_poll() instead of relying on the signal that
we get from the scan results.

BUG=b:111071506
TEST=Modify the driver to lie about the RSSI with the patch below.
Then, setup two APs with the same SSID: one on channel 1 and one on
channel 6 and check if we keep bouncing between the two APs.

diff --git a/drivers/net/wireless/intel/iwlwifi/mvm/rx.c b/drivers/net/wireless/intel/iwlwifi/mvm/rx.c
index 593b0ed..59296e1 100644
--- a/drivers/net/wireless/intel/iwlwifi/mvm/rx.c
+++ b/drivers/net/wireless/intel/iwlwifi/mvm/rx.c
@@ -168,11 +168,18 @@ static void iwl_mvm_get_signal_strength(struct iwl_mvm *mvm,
        energy_c = (val & IWL_RX_INFO_ENERGY_ANT_C_MSK) >>
                                                IWL_RX_INFO_ENERGY_ANT_C_POS;
        energy_c = energy_c ? -energy_c : S8_MIN;
+
+       if (le16_to_cpu(phy_info->channel) == 1)
+               energy_a -= 30;
+
+       if (le16_to_cpu(phy_info->channel) == 6)
+               energy_b -= 30;
+
        max_energy = max(energy_a, energy_b);
        max_energy = max(max_energy, energy_c);

-       IWL_DEBUG_STATS(mvm, "energy In A %d B %d C %d , and max %d\n",
-                       energy_a, energy_b, energy_c, max_energy);
+       IWL_DEBUG_STATS(mvm, "energy In A %d B %d C %d , and max %d  channel %d\n",
+                       energy_a, energy_b, energy_c, max_energy, le16_to_cpu(phy_info->channel));

        rx_status->signal = max_energy;
        rx_status->chains = (le16_to_cpu(phy_info->phy_flags) &

Change-Id: I21090246c6943c9ba57f2961c013246b835daa3a
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1201583
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>