BACKPORT: Update throughput estimate for the current BSS based on signal poll
We saw that on certain platforms in certain places we keep switching
between two APs and eventually get the same RSSI. Debugging showed that
we have a very big difference between the two antennas.

Ant A can hear AP A very well (-60) but AP B very bad (-80)
Ant B can hear AP B very well (-60) but AP A very bad (-80)

When the device associates to AP A, it'll learn to use Ant A. If the
device uses one single antenna to receive the scan results, it may hear
the AP it is currently associated to on the second antenna and get bad
results. Because of that, the wpa_supplicant will roam to the other AP
and the same scenario will repeat itself:

Association to AP A (Ant A reports -60).
Scan on Ant A: AP A: -60, AP B: -80
Scan on Ant B: AP A: -80, AP A: -60 ==> ROAM.

Association to AP B (Ant B reports -60)
Scan on Ant A: AP A: -60, AP B: -80 ==> ROAM

Etc...

Improve this by querying the signal level of the current AP using
drv_signal_poll() instead of relying on the signal level that we get
from the scan results. Also update the throughput estimate based on the
likely more accurate values for the current association.

(cherry picked from commit 7e7b23e2290b1c0731f0ae04200ddef033334abb)

Conflicts from log changes not taken by upstream.

BUG=chromium:1031320
TEST=deployed to elm device, triggered a roam by walking around,
     observed the log line that starts with "Using poll to get..."

Change-Id: I1b2516c846260519cdb62ae4af3153c1e7b92edb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2079743
Reviewed-by: Neeraj Poojary <npoojary@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>