CHROMIUM: wpa_supplicant: Implement roam hysteresis
Increase the difficulty to roam based on the time of the last roam, and
back it off as time goes on. We don't need to roam immediately once we
get the roam indication; if the user is indeed moving or is in view of a
better AP, they will eventually roam as the signal gets stronger (or
time passes and we determine that the AP is indeed roam-worthy).

BUG=chromium:1014866,chromium:1049782
TEST=1) network_WiFi_RoamingDiagnostics (WIP test) showed significant
        improvements (around 30% reduced roams).
     2) Normal roaming behavior walking around the building.

Change-Id: Ie82f76151e7f51e05f2f2651cb9a202ccba867d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2079750
Reviewed-by: Neeraj Poojary <npoojary@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>