FROMLIST: Decrease cross-threshold roam difficulty
Decrease the roam difficulty when we roam across the short background
scan threshold (i.e. when the current BSS RSSI is below the threshold
and the selected BSS RSSI is above the threshold). This tends to happen
in normal roaming scenarios when we walk from the domain of one AP to
the domain of the other.

BUG=chromium:1014866,chromium:1049782
TEST=1) network_WiFi_RoamingDiagnostics (WIP test) showed no significant
        decrease in stickiness.
     2) Roaming behavior improved when walking around the building
	(fewer disassocs).

(am from https://patchwork.ozlabs.org/patch/1301986)

Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Change-Id: I254f1c0c410ceedbe7fee86ab568c8aad81ce421
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hostap/+/2088811
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>